[query] Remove old implementation of TableJoin
Provides an example and template for doing the rest.
